:books: Update getBytes test
