+++ packages/lisk-transactions/test/utils/get_transaction_bytes.ts
@@ -764,7 +764,6 @@ describe('getTransactionBytes module', () => {
 					senderId: defaultSenderId,
 					signature: defaultSignature,
 					id: defaultTransactionId,
-					signatures: [],
 				};
 				return Promise.resolve();
 			});
